fix: harden #Tool param rewrite and split FM showcase scenarios
Replace naive regex rewriting in InlineToolMacro with AST-position
rewrites so string literals are not corrupted, and add regression tests.
Extract Foundation Models capability scenarios into a dedicated file to
stop CapabilityShowcase.swift from absorbing more provider-specific bulk.
